Suppress platform-dependent build errors. Some warnings remain.
This is a mess of conditional compilation and conditional clauses
in Package.swift. It introduces the SKIPMODULE define that the
platform implementations use to entirely skip their contents if the
manifest says to do so on that platform. It hopes to be closer to DRY
than encoding platform logic in each source file.